+++ packages/renderer/src/lib/container/ContainerList.svelte
@@ -1,45 +1,45 @@
 <script lang="ts">
 import { onDestroy, onMount } from 'svelte';
-import { filtered, searchPattern, containersInfos } from '../stores/containers';
-import { viewsContributions } from '../stores/views';
-import { context } from '../stores/context';
+import { filtered, searchPattern, containersInfos } from '../../stores/containers';
+import { viewsContributions } from '../../stores/views';
+import { context } from '../../stores/context';
 
-import type { ContainerInfo } from '../../../main/src/plugin/api/container-info';
-import PodIcon from './images/PodIcon.svelte';
-import StatusIcon from './images/StatusIcon.svelte';
+import type { ContainerInfo } from '../../../../main/src/plugin/api/container-info';
+import PodIcon from '../images/PodIcon.svelte';
+import StatusIcon from '../images/StatusIcon.svelte';
 import { router } from 'tinro';
-import { ContainerGroupInfoTypeUI, type ContainerGroupInfoUI, type ContainerInfoUI } from './container/ContainerInfoUI';
-import ContainerActions from './container/ContainerActions.svelte';
-import PodActions from './pod/PodActions.svelte';
-import ContainerIcon from './images/ContainerIcon.svelte';
-import ContainerEmptyScreen from './container/ContainerEmptyScreen.svelte';
-import FilteredEmptyScreen from './ui/FilteredEmptyScreen.svelte';
-import Modal from './dialogs/Modal.svelte';
-import { ContainerUtils } from './container/container-utils';
-import { providerInfos } from '../stores/providers';
-import NoContainerEngineEmptyScreen from './image/NoContainerEngineEmptyScreen.svelte';
+import { ContainerGroupInfoTypeUI, type ContainerGroupInfoUI, type ContainerInfoUI } from './ContainerInfoUI';
+import ContainerActions from './ContainerActions.svelte';
+import PodActions from '../pod/PodActions.svelte';
+import ContainerIcon from '../images/ContainerIcon.svelte';
+import ContainerEmptyScreen from './ContainerEmptyScreen.svelte';
+import FilteredEmptyScreen from '../ui/FilteredEmptyScreen.svelte';
+import Modal from '../dialogs/Modal.svelte';
+import { ContainerUtils } from './container-utils';
+import { providerInfos } from '../../stores/providers';
+import NoContainerEngineEmptyScreen from '../image/NoContainerEngineEmptyScreen.svelte';
 import moment from 'moment';
 import { get, type Unsubscriber } from 'svelte/store';
-import NavPage from './ui/NavPage.svelte';
+import NavPage from '../ui/NavPage.svelte';
 import { faChevronDown, faChevronRight, faPlusCircle, faTrash } from '@fortawesome/free-solid-svg-icons';
 import Fa from 'svelte-fa';
-import ErrorMessage from './ui/ErrorMessage.svelte';
-import { podCreationHolder } from '../stores/creation-from-containers-store';
-import { podsInfos } from '../stores/pods';
-import KubePlayButton from './kube/KubePlayButton.svelte';
-import Prune from './engine/Prune.svelte';
-import type { EngineInfoUI } from './engine/EngineInfoUI';
-import { containerGroupsInfo } from '../stores/containerGroups';
-import Checkbox from './ui/Checkbox.svelte';
-import type { PodInfo } from '../../../main/src/plugin/api/pod-info';
-import { PodUtils } from '../lib/pod/pod-utils';
-import ComposeActions from './compose/ComposeActions.svelte';
-import { CONTAINER_LIST_VIEW } from './view/views';
-import type { ViewInfoUI } from '../../../main/src/plugin/api/view-info';
-import type { ContextUI } from './context/context';
-import Button from './ui/Button.svelte';
-import StateChange from './ui/StateChange.svelte';
-import SolidPodIcon from './images/SolidPodIcon.svelte';
+import ErrorMessage from '../ui/ErrorMessage.svelte';
+import { podCreationHolder } from '../../stores/creation-from-containers-store';
+import { podsInfos } from '../../stores/pods';
+import KubePlayButton from '../kube/KubePlayButton.svelte';
+import Prune from '../engine/Prune.svelte';
+import type { EngineInfoUI } from '../engine/EngineInfoUI';
+import { containerGroupsInfo } from '../../stores/containerGroups';
+import Checkbox from '../ui/Checkbox.svelte';
+import type { PodInfo } from '../../../../main/src/plugin/api/pod-info';
+import { PodUtils } from '../pod/pod-utils';
+import ComposeActions from '../compose/ComposeActions.svelte';
+import { CONTAINER_LIST_VIEW } from '../view/views';
+import type { ViewInfoUI } from '../../../../main/src/plugin/api/view-info';
+import type { ContextUI } from '../context/context';
+import Button from '../ui/Button.svelte';
+import StateChange from '../ui/StateChange.svelte';
+import SolidPodIcon from '../images/SolidPodIcon.svelte';
 
 const containerUtils = new ContainerUtils();
 let openChoiceModal = false;
