+++ packages/renderer/src/lib/image/ImagesList.spec.ts
@@ -22,9 +22,9 @@ import '@testing-library/jest-dom/vitest';
 import { beforeAll, test, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/svelte';
 import ImagesList from './ImagesList.svelte';
-import { imagesInfos } from '../stores/images';
+import { imagesInfos } from '../../stores/images';
 import { get } from 'svelte/store';
-import { providerInfos } from '../stores/providers';
+import { providerInfos } from '../../stores/providers';
 
 const listImagesMock = vi.fn();
 const getProviderInfosMock = vi.fn();
