+++ packages/renderer/src/lib/image/ImagesList.svelte
@@ -1,29 +1,29 @@
 <script lang="ts">
-import { filtered, searchPattern, imagesInfos } from '../stores/images';
+import { filtered, searchPattern, imagesInfos } from '../../stores/images';
 import { onDestroy, onMount } from 'svelte';
-import ImageEmptyScreen from './image/ImageEmptyScreen.svelte';
-import FilteredEmptyScreen from './ui/FilteredEmptyScreen.svelte';
+import ImageEmptyScreen from './ImageEmptyScreen.svelte';
+import FilteredEmptyScreen from '../ui/FilteredEmptyScreen.svelte';
 
 import { router } from 'tinro';
-import type { ImageInfoUI } from './image/ImageInfoUI';
-import ImageActions from './image/ImageActions.svelte';
-import type { ImageInfo } from '../../../main/src/plugin/api/image-info';
-import NoContainerEngineEmptyScreen from './image/NoContainerEngineEmptyScreen.svelte';
-import { providerInfos } from '../stores/providers';
-import PushImageModal from './image/PushImageModal.svelte';
-import RenameImageModal from './image/RenameImageModal.svelte';
-import { ImageUtils } from './image/image-utils';
-import NavPage from './ui/NavPage.svelte';
-import ImageIcon from './images/ImageIcon.svelte';
-import StatusIcon from './images/StatusIcon.svelte';
+import type { ImageInfoUI } from './ImageInfoUI';
+import ImageActions from './ImageActions.svelte';
+import type { ImageInfo } from '../../../../main/src/plugin/api/image-info';
+import NoContainerEngineEmptyScreen from './NoContainerEngineEmptyScreen.svelte';
+import { providerInfos } from '../../stores/providers';
+import PushImageModal from './PushImageModal.svelte';
+import RenameImageModal from './RenameImageModal.svelte';
+import { ImageUtils } from './image-utils';
+import NavPage from '../ui/NavPage.svelte';
+import ImageIcon from '../images/ImageIcon.svelte';
+import StatusIcon from '../images/StatusIcon.svelte';
 import type { Unsubscriber } from 'svelte/store';
-import { containersInfos } from '../stores/containers';
-import type { ContainerInfo } from '../../../main/src/plugin/api/container-info';
+import { containersInfos } from '../../stores/containers';
+import type { ContainerInfo } from '../../../../main/src/plugin/api/container-info';
 import moment from 'moment';
-import Prune from './engine/Prune.svelte';
-import type { EngineInfoUI } from './engine/EngineInfoUI';
-import Checkbox from './ui/Checkbox.svelte';
-import Button from './ui/Button.svelte';
+import Prune from '../engine/Prune.svelte';
+import type { EngineInfoUI } from '../engine/EngineInfoUI';
+import Checkbox from '../ui/Checkbox.svelte';
+import Button from '../ui/Button.svelte';
 import { faArrowCircleDown, faCube, faTrash } from '@fortawesome/free-solid-svg-icons';
 
 export let searchTerm = '';
